Use formatTime for time-related functions
